Refactor importer namespace to remove data type
This is due to TP importer populating daily calls
and where did you hear data.
