+++ app/jobs/import_tp_data_job.rb
@@ -4,6 +4,6 @@ class ImportTPDataJob < ActiveJob::Base
   queue_as :default
 
   def perform
-    Importers::DailyCalls::TP::Importer.new.import
+    Importers::TP::Importer.new.import
   end
 end
