+++ lib/importers/tp.rb
@@ -1,0 +1,8 @@
+module Importers
+  module TP
+    autoload :CallRecord, 'importers/tp/call_record'
+    autoload :Importer, 'importers/tp/importer'
+    autoload :Retriever, 'importers/tp/retriever'
+    autoload :Saver, 'importers/tp/saver'
+  end
+end
