+++ lib/importers/twilio.rb
@@ -1,0 +1,8 @@
+module Importers
+  module Twilio
+    autoload :CallRecord, 'importers/twilio/call_record'
+    autoload :Importer, 'importers/twilio/importer'
+    autoload :Retriever, 'importers/twilio/retriever'
+    autoload :Saver, 'importers/twilio/saver'
+  end
+end
